+++ src/lib/components/SlotView/MemberInput/FieldInput.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { getTsTypeFromResoniteType, isNullableType } from "@coin/resonitelink-ts";
-    import type {boolean2, boolean3, boolean4, color, colorX, Field, FieldGeneric, number2, number2x2, number3, number3x3, number4, number4x4, numberQ} from "@coin/resonitelink-ts"
+    import type {boolean2, boolean3, boolean4, color, colorX, Field, FieldGeneric, number2, number2x2, number3, number3x3, number4, number4x4, numberQ, rect} from "@coin/resonitelink-ts"
     import FieldNumber from "./Fields/FieldNumber.svelte";
     import FieldNumber3 from "./Fields/FieldNumber3.svelte";
     import FieldNumber2 from "./Fields/FieldNumber2.svelte";
@@ -19,6 +19,7 @@
     import FieldNumber2x2 from "./Fields/FieldNumber2x2.svelte";
     import FieldNumber3x3 from "./Fields/FieldNumber3x3.svelte";
     import FieldNumber4x4 from "./Fields/FieldNumber4x4.svelte";
+    import FieldRect from "./Fields/FieldRect.svelte";
     const { data, changeField }: { data: Field, changeField: (data: Field) => void } = $props()
 
     const tsType = $derived(getTsTypeFromResoniteType(data.$type))
@@ -66,6 +67,8 @@
     <FieldStringNullable data={data as FieldGeneric<string, string|null>} {changeField}/>
 {:else if tsType === "string"}
     <FieldString data={data as FieldGeneric<string, string>} {changeField}/>
+{:else if tsType === "rect"}
+    <FieldRect data={data as FieldGeneric<string, rect>} {changeField}/>
 {:else}
     <span class="type">{tsType}</span>
 {/if}